Write up ported class
