+++ FirebaseRemoteConfig/Tests/Unit/RCNRemoteConfigTest.m
@@ -25,7 +25,6 @@
 #import "FirebaseRemoteConfig/Sources/Private/RCNConfigFetch.h"
 #import "FirebaseRemoteConfig/Sources/Public/FirebaseRemoteConfig/FIRRemoteConfig.h"
 #import "FirebaseRemoteConfig/Sources/RCNConfigConstants.h"
-#import "FirebaseRemoteConfig/Sources/RCNConfigExperiment.h"
 #import "FirebaseRemoteConfig/Sources/RCNConfigRealtime.h"
 
 #import "FirebaseRemoteConfig/Tests/Unit/RCNTestUtilities.h"
